+++ ec-project-api/Models/products/ProductVariant.cs
@@ -11,7 +11,7 @@ namespace ec_project_api.Models {
         public int ProductId { get; set; }
 
         [Column("size_id")]
-        public byte SizeId { get; set; }
+        public short SizeId { get; set; }
 
         [Required]
         [StringLength(50)]
